requirements.txt
# ============================================================
#  Cassandre 🔮 — Dépendances
#  Requis : Python 3.8+ (Recommandé : 3.10 - 3.12)
# ============================================================

# Core
psutil==5.9.8
pandas>=2.0.0
SQLAlchemy
python-dotenv>=1.0.0

# Data & Finance
yfinance==0.2.36
ccxt==4.5.48
ta==0.11.0

# UI & Viz
gradio==4.44.1
plotly>=5.18.0
pywebview; sys_platform == "win32" or sys_platform == "darwin"

# Compatibility Fixes
Jinja2>=3.1.5           # Python 3.14 : dict key hashability fix dans LRUCache
huggingface-hub<0.26.0
audioop-lts>=0.2.1; python_version >= "3.13"   # audioop retiré en Python 3.13+

# IA & Intelligence
scikit-learn
numpy

